Add dummy data to map's sidebar
