+++ utmap-client/src/components/pages/MapPage.js
@@ -4,6 +4,7 @@ import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
 import SideBar from "./SideBar";
 import clsx from 'clsx';
+import exampleEvents from '../data/EventData';//Temporary dummy data
   
 const useStyles = makeStyles(theme => ({
     hide: {
@@ -37,7 +38,7 @@ function MapPage() {
           >
           <MenuIcon />
         </IconButton>
-        <SideBar open={openDrawer} onClose={handleDrawerClose}></SideBar>
+        <SideBar open={openDrawer} onClose={handleDrawerClose} events={exampleEvents}/>
       </div>
       <p>Map</p>
     </>
